requirements.txt
pytest
pytest-cov
python-dotenv
pandas
openpyxl
pytest-xdist
pytest-rerunfailures
requests
mysql-connector-python
mysql
selenium
allure-pytest